requirements.txt
openai>=1.0.0
langchain>=0.1.0
faiss-cpu
PyMuPDF
streamlit
python-dotenv
langchain-community